Refactor LeetCode integration: remove old API handler and implement new server-side proxy for fetching user statistics
