--- src/routes/api/leetcode.ts
@@ -1,16 +1,0 @@
-import type { RequestHandler } from '@sveltejs/kit';
-
-export const GET: RequestHandler = async ({ url }) => {
-    const response = await fetch('https://leetcode.com/graphql', {
-        headers: {
-            'Content-Type': 'application/json'
-        }
-    });
-
-    const data = await response.json();
-    return new Response(JSON.stringify(data), {
-        headers: {
-            'Content-Type': 'application/json'
-        }
-    });
-};
